Prevent accidental deletion of workspace in multi-stage builds

+++ pkg/executor/build.go
@@ -492,8 +492,10 @@ func DoBuild(opts *config.KanikoOptions) (v1.Image, error) {
 		}
 
 		// Delete the filesystem
-		if err := util.DeleteFilesystem(); err != nil {
-			return nil, err
+		if opts.Cleanup {
+			if err := util.DeleteFilesystem(); err != nil {
+				return nil, err
+			}
 		}
 	}
